Initial support for a LocalID service
