+++ ingest-src/src/main/java/org/cdlib/mrt/ingest/handlers/HandlerTransfer.java
@@ -60,6 +60,7 @@ import org.cdlib.mrt.ingest.IngestRequest;
 import org.cdlib.mrt.ingest.JobState;
 import org.cdlib.mrt.ingest.ProfileState;
 import org.cdlib.mrt.ingest.StoreNode;
+import org.cdlib.mrt.ingest.utility.LocalIDUtil;
 import org.cdlib.mrt.ingest.utility.ProfileUtil;
 import org.cdlib.mrt.ingest.utility.StorageUtil;
 import org.cdlib.mrt.ingest.utility.TExceptionResponse;
@@ -144,17 +145,6 @@ public class HandlerTransfer extends Handler<JobState>
 	            if (DEBUG) System.out.println("[debug] " + MESSAGE + " delete file found: " + deleteFile.getName());
   	    	    formData.add("delete", processDeleteFile(deleteFile));
 		}
-
-	    }
-	    try {
-		if (jobState.getLocalID().getValue().contains("(:unas)")) {
-                    if (DEBUG) System.out.println("[debug] " + MESSAGE + "No Local ID specified for object");
-                    throw new Exception("");	// (:unas) is equivalent to no localID
-		}
-
-  	        formData.add("local-identifier", jobState.getLocalID().getValue());
-  	        formData.add("local-context", profileState.getOwner());	
-	    } catch (Exception e) {
 	    }
 
 	    // make service request
@@ -190,6 +180,26 @@ public class HandlerTransfer extends Handler<JobState>
 
 	    jobState.setCompletionDate(new DateState(DateUtil.getCurrentDate()));
 	    jobState.setVersionID(getVersionID(clientResponse.getEntity(String.class)));
+
+	    // Update the LocalID db
+	    try {
+		if (jobState.getLocalID().getValue().contains("(:unas)")) {
+                    if (DEBUG) System.out.println("[debug] " + MESSAGE + "No Local ID found.");
+                    throw new Exception("");	// (:unas) is equivalent to no localID
+		}
+
+		try {
+                   if (DEBUG) System.out.println("[debug] " + MESSAGE + "Updating LocalID db pid: " + 
+			jobState.getPrimaryID().getValue() + " lid: " + jobState.getLocalID().getValue());
+		   LocalIDUtil.addLocalID(profileState, jobState.getPrimaryID().getValue(), jobState.getLocalID().getValue());
+		} catch (Exception le) {
+		   le.printStackTrace();
+                   System.err.println("[error] " + MESSAGE + "failed to update LocalID db: " + le.getMessage());
+		   throw le;
+		}
+	    } catch (Exception e) {
+		// no local ID
+	    }
 
 	    return new HandlerResult(true, "SUCCESS: transfer", clientResponse.getStatus());
 	} catch (TException te) {
