+++ ingest-src/src/main/java/org/cdlib/mrt/ingest/utility/StorageUtil.java
@@ -258,281 +258,4 @@ public class StorageUtil
         }
     }
 
-
-    public static String fetchPrimaryID(ProfileState profileState, String localID)
-        throws TException
-    {
-
-        ClientResponse clientResponse = null;
-        try {
-	    // First lets fetch from our localID DB on Storage
-	    String primaryID = null;
-            StoreNode storeNode = profileState.getTargetStorage();
-
-            // build REST url
-            String url = storeNode.getStorageLink().toString() + "/primary/" + storeNode.getNodeID() + "/" + 
-			URLEncoder.encode(profileState.getOwner(), "utf-8") + "/" +
-			URLEncoder.encode(localID,  "utf-8") + "?t=xml";
-            if (DEBUG) System.out.println("[debug] PrimaryID fetch URL: " + url);
-            Client client = Client.create();    // reuse?  creation is expensive
-
-	    /* fix client timeout problem */
-	    client.setConnectTimeout(new Integer(STORAGE_CONNECT_TIMEOUT_SHORT));
-	    client.setReadTimeout(new Integer(STORAGE_READ_TIMEOUT_SHORT));
-
-            WebResource webResource = client.resource(url);
-
-            // make service request
-            clientResponse = webResource.get(ClientResponse.class);
-	    int status = clientResponse.getStatus();
-   	    String response = null;
- 
-	    if (status != 200) {
-                try {
-                    // most likely exception
-                    // can only call once, as stream is not reset
-                    TExceptionResponse.REQUEST_INVALID tExceptionResponse = clientResponse.getEntity(TExceptionResponse.REQUEST_INVALID.class);
-                    throw new TException.REQUEST_INVALID(tExceptionResponse.getError());
-                } catch (TException te) {
-                    throw te;
-                } catch (Exception e) {
-                    // let's report something
-                    throw new TException.EXTERNAL_SERVICE_UNAVAILABLE("[error] " + NAME + ": storage service: " + url);
-                }
-	    } else {
-   	        response = clientResponse.getEntity(String.class);
-	    }
-
-            DocumentBuilderFactory domFactory = DocumentBuilderFactory.newInstance();
-            domFactory.setNamespaceAware(true);
-            domFactory.setExpandEntityReferences(true);
-
-            DocumentBuilder builder = domFactory.newDocumentBuilder();
-            builder.setErrorHandler(null);
-            Document document = builder.parse(new ByteArrayInputStream(response.getBytes("UTF-8")));
-            XPath xpath = XPathFactory.newInstance().newXPath();
-            XPathExpression expr = xpath.compile("//*[local-name()='primaryIdentifier']");
-
-            String xpathS = (String) expr.evaluate(document);
-            if (StringUtil.isNotEmpty(xpathS)) {
-                if (DEBUG) System.out.println("[debug] primary ID: " + xpathS);
-                primaryID = xpathS;
-            } else {
-                if (DEBUG) System.out.println("[debug] Can not determine primary ID from storage DB");
-		primaryID = null;
-            }
-
-	    // Next lets try EZID if local looks like a DOI
-	    // note: EZID service is SSL so don't use Jersey client
-	    if (primaryID == null && localID.toLowerCase().startsWith("doi:")) {
-                // build REST url
-                url = profileState.getObjectMinterURL().toString();
-                url = url.replaceFirst("/shoulder.*", "/id/") + localID.toLowerCase();
-
-                System.out.println("[info] " + MESSAGE + "Searching for shadowedBy ID: " + url);
-
-		DefaultHttpClient httpClient = new DefaultHttpClient();
-                httpClient = wrapClient(httpClient);
-
-		// authenticate
-		String misc = null;
-		if ((misc = profileState.getMisc()) == null) {
-                    System.err.println("[warning] " + MESSAGE + "EZID credentials not found.");
-                    throw new TException.GENERAL_EXCEPTION("EZID credentials not found.");
-		}
-
-
-		String[] auth = misc.split(":");
-		Credentials credentials = new UsernamePasswordCredentials( auth[0], auth[1] );
-		httpClient.getCredentialsProvider().setCredentials(new AuthScope(AuthScope.ANY_HOST, AuthScope.ANY_PORT), credentials);
-		HttpGet httpGet = null;
-
-		try {
-		    httpGet = new HttpGet(url);
-		} catch (java.lang.IllegalArgumentException iae) {
-		    throw new TException.INVALID_OR_MISSING_PARM("Target hostname or primary ID not valid: " + url);
-		}
-		httpGet.addHeader("Accept", "text/plain");
-
-
-		String responseBody = null;
-		HttpResponse httpResponse = null;
-
-		try {
-		    httpResponse = httpClient.execute(httpGet);
-		    responseBody = StringUtil.streamToString(httpResponse.getEntity().getContent(), "UTF-8");
-		} catch (HttpHostConnectException hhce) {
-		    throw new TException.EXTERNAL_SERVICE_UNAVAILABLE("error in connecting to host: " + url);
-		} catch (org.apache.http.client.HttpResponseException hre) {
-		    System.err.println("[error] " + MESSAGE + "request failed with status code: " + hre.getStatusCode());
-		    System.err.println("[error] " + MESSAGE + "request failed with status message: " + hre.getMessage());
-		    responseBody = "failed";
-		}
-
-		status = httpResponse.getStatusLine().getStatusCode();
-
-                if (status == 400) {
-		    // DOI not registered with EZID
-		    if (DEBUG) System.out.println("[warn] " + MESSAGE + " DOI local ID is NOT registered with EZID: " + localID);
-		    primaryID = null;
-		} else if (status == 200) {
-		    // Found DOI in EZID
-		    String coowners = "";
-		    String owner = "";
-		    // extract ARK listed as "_shadowedBy"
-		    String[] responseSplit = responseBody.split("\n");
-		    for (String element : responseSplit) {
-			if (element.startsWith("_shadowedby:")) {
-			    primaryID = element.substring(element.indexOf(":") + 1).trim();
-			    if (DEBUG) System.out.println("[debug] " + MESSAGE + " Found DOI in EZID, shadow ark is: " + primaryID);
-			}
-			if (element.startsWith("_coowners:")) {
-			    coowners = element.substring(element.indexOf(":") + 1).trim();
-			}
-			if (element.startsWith("_owner:")) {
-			    owner = element.substring(element.indexOf(":") + 1).trim();
-			}
-		    }
-		    // Are we permitted to update this DOI?
-		    if ((! coowners.contains("merritt")) && (! owner.contains("merritt"))) {
-                        throw new TException.REQUEST_INVALID("[error] " + NAME + ": Merritt not authorized to update DOI: " + url);
-		    }
-                } else {
-		    // error
-                    throw new TException.EXTERNAL_SERVICE_UNAVAILABLE("[error] " + NAME + ": EZID service: " + url);
-                }
-            }
-
-            return primaryID;
-
-        } catch (TException te) {
-	    throw te;
-        } catch (com.sun.jersey.api.client.ClientHandlerException che) {
-            che.printStackTrace();
-            String msg = "[error] " + MESSAGE + "Could not connect to Storage service: " 
-	        + profileState.getTargetStorage().getStorageLink().toString() + " --- " + che.getMessage();
-            throw new TException.EXTERNAL_SERVICE_UNAVAILABLE(msg);
-        } catch (Exception e) {
-            e.printStackTrace();
-            String msg = "[error] " + MESSAGE + "failed to map localID. " + e.getMessage();
-            throw new TException.GENERAL_EXCEPTION(msg);
-        } finally {
-            try {
-            } catch (Exception e) {}
-        }
-    }
-
-    public static String fetchLocalID(ProfileState profileState, String primaryID)
-        throws TException
-    {
-
-        ClientResponse clientResponse = null;
-        try {
-
-	    //Vector localIDs = null;
-	    String localIDs = null;
-            StoreNode storeNode = profileState.getTargetStorage();
-
-            // build REST url
-            String url = storeNode.getStorageLink().toString() + "/local/" + storeNode.getNodeID() + "/" + 
-			URLEncoder.encode(primaryID,  "utf-8") + "?t=xml";
-            if (DEBUG) System.out.println("[debug] LocalID fetch URL from storage db: " + url);
-            Client client = Client.create();    // reuse?  creation is expensive
-
-            /* fix client timeout problem */
-            client.setConnectTimeout(new Integer(STORAGE_CONNECT_TIMEOUT_SHORT));
-            client.setReadTimeout(new Integer(STORAGE_READ_TIMEOUT_SHORT));
-
-            WebResource webResource = client.resource(url);
-
-            // make service request
-            clientResponse = webResource.get(ClientResponse.class);
-	    int status = clientResponse.getStatus();
-   	    String response = null;
- 
-	    if (status != 200) {
-                try {
-                    // most likely exception
-                    // can only call once, as stream is not reset
-                    TExceptionResponse.REQUEST_INVALID tExceptionResponse = clientResponse.getEntity(TExceptionResponse.REQUEST_INVALID.class);
-                    throw new TException.REQUEST_INVALID(tExceptionResponse.getError());
-                } catch (TException te) {
-                    throw te;
-                } catch (Exception e) {
-                    // let's report something
-                    throw new TException.EXTERNAL_SERVICE_UNAVAILABLE("[error] " + NAME + ": storage service: " + url);
-                }
-	    } else {
-   	        response = clientResponse.getEntity(String.class);
-	    }
-
-            DocumentBuilderFactory domFactory = DocumentBuilderFactory.newInstance();
-            domFactory.setNamespaceAware(true);
-            domFactory.setExpandEntityReferences(true);
-
-            DocumentBuilder builder = domFactory.newDocumentBuilder();
-            builder.setErrorHandler(null);
-            Document document = builder.parse(new ByteArrayInputStream(response.getBytes("UTF-8")));
-            XPath xpath = XPathFactory.newInstance().newXPath();
-            XPathExpression expr = xpath.compile("//*[local-name()='localID']");
-
- 	    NodeList nl = (NodeList) expr.evaluate(document, XPathConstants.NODESET);
-    	    int i = 0;
-    	    for (i = 0; i < nl.getLength(); i++) {
-		Node n = nl.item(i);
-		String localid = n.getTextContent();
-		if (StringUtil.isNotEmpty(localid)) {
-                    if (DEBUG) System.out.println("[debug] LocalID found (storage db): " + localid);
-		    if (localIDs == null) localIDs = localid;
-		    else localIDs += "; " + localid;
-		}
-    	    }
-
-	    if (StringUtil.isEmpty(localIDs)) if (DEBUG) System.out.println("[debug] Can not determine local ID");
-            return localIDs;
-
-        } catch (TException te) {
-	    throw te;
-        } catch (com.sun.jersey.api.client.ClientHandlerException che) {
-            che.printStackTrace();
-            String msg = "[error] " + MESSAGE + "Could not connect to Storage service: " 
-	        + profileState.getTargetStorage().getStorageLink().toString() + " --- " + che.getMessage();
-            throw new TException.EXTERNAL_SERVICE_UNAVAILABLE(msg);
-        } catch (Exception e) {
-            e.printStackTrace();
-            String msg = "[error] " + MESSAGE + "failed to map primaryID to localID. " + e.getMessage();
-            throw new TException.GENERAL_EXCEPTION(msg);
-        } finally {
-            try {
-            } catch (Exception e) {}
-        }
-    }
-
-    // http://theskeleton.wordpress.com/2010/07/24/
-        // avoiding-the-javax-net-ssl-sslpeerunverifiedexception-peer-not-authenticated-with-httpclient/
-    public static DefaultHttpClient wrapClient(DefaultHttpClient base) {
-        try {
-            SSLContext ctx = SSLContext.getInstance("TLS");
-            X509TrustManager tm = new X509TrustManager() {
-
-                public void checkClientTrusted(X509Certificate[] xcs, String string) throws CertificateException { }
-
-                public void checkServerTrusted(X509Certificate[] xcs, String string) throws CertificateException { }
-
-                public X509Certificate[] getAcceptedIssuers() {
-                    return null;
-                }
-            };
-            ctx.init(null, new TrustManager[]{tm}, null);
-            SSLSocketFactory ssf = new SSLSocketFactory(ctx);
-            ssf.setHostnameVerifier(SSLSocketFactory.ALLOW_ALL_HOSTNAME_VERIFIER);
-            ClientConnectionManager ccm = base.getConnectionManager();
-            SchemeRegistry sr = ccm.getSchemeRegistry();
-            sr.register(new Scheme("https", ssf, 443));
-            return new DefaultHttpClient(ccm, base.getParams());
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            return null;
-        }
-    }
 }
